RAVE 0.424
- Afegida finestra que informa de la pre-carrega de la libvlc (abans de que surti la finesrta principal)
- Solucionat bug al utilitzar consultes Varg... ja que si li pases un string NULL el comportament es indeterminat... S'ha eliminat la funcio varg i ara es fara tot concatenan el string amb els seus operadors.
- La funció d'actualitzar l'arbre ara espera a que el Thread del analisis hagi acabat abans de tornar a començar tot el proces

+++ Reproductor de Audio y Video Extendido/DMenuEx.cpp
@@ -522,7 +522,7 @@ namespace DWL {
 		// Obtengo la posición del menu resaltado dentro del vector
 		int MenuPos = -1;
 		if (_MenuResaltado != NULL) {
-			for (MenuPos = 0; MenuPos < _Menus.size(); MenuPos++) {
+			for (MenuPos = 0; MenuPos < static_cast<int>(_Menus.size()); MenuPos++) {
 				if (_Menus[MenuPos] == _MenuResaltado) {
 					break;
 				}
